разбивает один файл на модули
